Add rebasing for trusted-cage object pointers.
Pointers to trusted-cage objects are embedded in
code as absolute  addresses via the corresponding
relocation type because the trusted-cage base
isn't readily available at the runtime sites
where we need to decompress a pointer to
a trusted object.

During cloning, when the trusted-cage base
changes, we rebase all such pointers. This path is
rare, so the impact on clone time is negligible.