Add cloning routines for the JavaScript Dispatch Handles Table (JDT).
Each JDT handle consists of two pointers:
an entry-point pointer and a pointer to the
corresponding code object. Entry points reside
in the code cage, while code-object pointers
target either the trusted cage
or the heap pointer cage. To clone the JDT, copy
the segment data and rebase all pointers to the
corresponding cages in the clone.